+++ UI/ViewModels/MainWindowViewModel.cs
@@ -14,7 +14,8 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
 {
     internal class MainWindowViewModel : NotifyBase
     {
-        private const string ButtonTooltipMessage = "Family, Title and Indexes are required fields.";
+        private const string SaveFileButtonTooltipMessage = "Family, Title and Indexes are required fields.";
+        private const string OpenOutputFolderButtonTooltipMessage = "Save a file first.";
 
         private readonly UserSelectionFileManager userSelectionFileManager;
         private string family;
@@ -22,9 +23,11 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
         private string title;
         private string comment;
         private string indexesInput;
-        private string buttonTooltip;
+        private string saveFileButtonTooltip;
+        private string openOutputFolderButtonTooltip;
         private string versionLabel;
         private string latestLoadSaveFolderPath;
+        private string latestFileSavedPath;
 
         public string Family
         {
@@ -39,7 +42,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                 {
                     family = value;
                     RaisePropertyChanged();
-                    UpdateCanExecute();
+                    UpdateCanExecuteSaveFileCommand();
                 }
             }
         }
@@ -57,7 +60,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                 {
                     model = value;
                     RaisePropertyChanged();
-                    UpdateCanExecute();
+                    UpdateCanExecuteSaveFileCommand();
                 }
             }
         }
@@ -75,7 +78,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                 {
                     title = value;
                     RaisePropertyChanged();
-                    UpdateCanExecute();
+                    UpdateCanExecuteSaveFileCommand();
                 }
             }
         }
@@ -93,7 +96,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                 {
                     comment = value;
                     RaisePropertyChanged();
-                    UpdateCanExecute();
+                    UpdateCanExecuteSaveFileCommand();
                 }
             }
         }
@@ -111,23 +114,40 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                 {
                     indexesInput = value;
                     RaisePropertyChanged();
-                    UpdateCanExecute();
+                    UpdateCanExecuteSaveFileCommand();
                 }
             }
         }
 
-        public string ButtonTooltip
+        public string SaveFileButtonTooltip
         {
             get
             {
-                return buttonTooltip;
+                return saveFileButtonTooltip;
             }
 
             set
             {
-                if (value != buttonTooltip)
+                if (value != saveFileButtonTooltip)
                 {
-                    buttonTooltip = value;
+                    saveFileButtonTooltip = value;
+                    RaisePropertyChanged();
+                }
+            }
+        }
+
+        public string OpenOutputFolderButtonTooltip
+        {
+            get
+            {
+                return openOutputFolderButtonTooltip;
+            }
+
+            set
+            {
+                if (value != openOutputFolderButtonTooltip)
+                {
+                    openOutputFolderButtonTooltip = value;
                     RaisePropertyChanged();
                 }
             }
@@ -152,6 +172,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
 
         public RelayCommand LoadFileCommand { get; private set; }
         public RelayCommand SaveFileCommand { get; private set; }
+        public RelayCommand OpenOutputFolderCommand { get; private set; }
         public RelayCommand ClearFormCommand { get; private set; }
 
         public MainWindowViewModel(UserSelectionFileManager userSelectionFileManager)
@@ -161,6 +182,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
             this.userSelectionFileManager = userSelectionFileManager;
             LoadFileCommand = new RelayCommand(o => ExecuteLoadFileCommand());
             SaveFileCommand = new RelayCommand(o => ExecuteSaveFileCommand(), o => CanExecuteSaveFileCommand());
+            OpenOutputFolderCommand = new RelayCommand(o => ExecuteOpenOutputFolderCommand(), o => CanExecuteOpenOutputFolderCommand());
             ClearFormCommand = new RelayCommand(o => ExecuteClearFormCommand());
 
             VersionLabel = GetAssemblyVersion();
@@ -175,8 +197,12 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
             Title = string.Empty;
             Comment = string.Empty;
             IndexesInput = string.Empty;
-            ButtonTooltip = ButtonTooltipMessage;
+            SaveFileButtonTooltip = SaveFileButtonTooltipMessage;
+            OpenOutputFolderButtonTooltip = OpenOutputFolderButtonTooltipMessage;
             latestLoadSaveFolderPath = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            latestFileSavedPath = null;
+            UpdateCanExecuteSaveFileCommand();
+            UpdateCanExecuteOpenOutputFolderCommand();
         }
 
         private void ExecuteLoadFileCommand()
@@ -187,6 +213,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
 
             if (loadFileDialog.ShowDialog() == true)
             {
+                InitializeForm();
                 latestLoadSaveFolderPath = Path.GetDirectoryName(loadFileDialog.FileName);
                 var data = userSelectionFileManager.LoadFile(loadFileDialog.FileName);
 
@@ -200,11 +227,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                     Model = data.Model;
                     Title = data.Title;
                     Comment = data.Comment;
-
-                    foreach (var item in data.Indexes)
-                    {
-                        IndexesInput += item + Environment.NewLine;
-                    }
+                    IndexesInput = IndexesFormatter.ToString(data.Indexes);
                 }
             }
         }
@@ -212,7 +235,7 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
         private bool CanExecuteSaveFileCommand()
         {
             var fieldsAreValid = FieldsAreValid();
-            UpdateButtonTooltip(fieldsAreValid);
+            UpdateSaveFileButtonTooltip(fieldsAreValid);
             return fieldsAreValid;
         }
 
@@ -236,17 +259,38 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
                         Model = Model,
                         Title = Title,
                         Comment = Comment,
-                        Indexes = ParseIndexesList()
+                        Indexes = IndexesFormatter.ToList(IndexesInput)
                     };
-                    userSelectionFileManager.CreateFile(data, saveFileDialog.FileName);
 
-                    OpenSavingFolder(folderPath);
+                    var operationSuccessful = userSelectionFileManager.CreateFile(data, saveFileDialog.FileName);
+
+                    if (operationSuccessful)
+                    {
+                        latestFileSavedPath = saveFileDialog.FileName;
+                        UpdateCanExecuteOpenOutputFolderCommand();
+                    }
+                    else
+                    {
+                        MessageBox.Show("An error ocurred while saving the file!", "Alert", MessageBoxButton.OK, MessageBoxImage.Error);
+                    }
                 }
                 else
                 {
                     MessageBox.Show("Location invalid!", "Alert", MessageBoxButton.OK, MessageBoxImage.Error);
                 }
             }
+        }
+
+        private bool CanExecuteOpenOutputFolderCommand()
+        {
+            var canOpen = !string.IsNullOrWhiteSpace(latestFileSavedPath);
+            UpdateOpenOutputFolderButtonTooltip(canOpen);
+            return canOpen;
+        }
+
+        private void ExecuteOpenOutputFolderCommand()
+        {
+            OpenSavingFolder(latestFileSavedPath);
         }
 
         private void ExecuteClearFormCommand()
@@ -261,46 +305,34 @@ namespace SysmacStudioParameterEditorUserSelectionFileMaker.UI.ViewModels
 
         private bool IndexesListIsValid()
         {
-            return ParseIndexesList().Count() > 0;
+            return IndexesFormatter.ToList(IndexesInput).Count() > 0;
         }
 
-        private IList<string> ParseIndexesList()
+        private void UpdateCanExecuteOpenOutputFolderCommand()
         {
-            var result = new List<string>();
-
-            var lines = IndexesInput.Split('\n');
-
-            foreach (var line in lines)
-            {
-                var indexes = line.Split(' ');
-
-                foreach (var index in indexes)
-                {
-                    if (!string.IsNullOrWhiteSpace(index))
-                    {
-                        result.Add(index.Trim());
-                    }
-                }
-            }
-
-            return result;
+            OpenOutputFolderCommand.RaiseCanExecuteChanged();
         }
 
-        private void UpdateCanExecute()
+        private void UpdateCanExecuteSaveFileCommand()
         {
             SaveFileCommand.RaiseCanExecuteChanged();
         }
 
-        private void UpdateButtonTooltip(bool hideTooltip)
+        private void UpdateOpenOutputFolderButtonTooltip(bool hideTooltip)
         {
-            ButtonTooltip = hideTooltip ? string.Empty : ButtonTooltipMessage;
+            OpenOutputFolderButtonTooltip = hideTooltip ? string.Empty : OpenOutputFolderButtonTooltipMessage;
         }
 
-        private void OpenSavingFolder(string folderPath)
+        private void UpdateSaveFileButtonTooltip(bool hideTooltip)
+        {
+            SaveFileButtonTooltip = hideTooltip ? string.Empty : SaveFileButtonTooltipMessage;
+        }
+
+        private void OpenSavingFolder(string filePath)
         {
             ProcessStartInfo startInfo = new ProcessStartInfo
             {
-                Arguments = folderPath,
+                Arguments = $"/select,\"{filePath}\"",
                 FileName = "explorer.exe"
             };