+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024 - 03 - 14
+
+### Added
+
+- Indexes can be separated in the same line by ',' and ';' as well as ' '. Empty entries will be removed.
+- Latest saved file is selected upon output folder opening.
+
+### Changed
+
+- Output folder opening with a button instead of automatically after saving file.
+
 ## [1.0.1] - 2024 - 03 - 11
 
 ### Fixed